Add keyboard-driven organization mode picker
