+++ extensions/folder-organizer/src/clean-downloads-folder.tsx
@@ -1,46 +1,31 @@
-import { showToast, Toast, Alert, confirmAlert } from "@raycast/api";
-import { showInFinder } from "@raycast/api";
+import { closeMainWindow, showToast, Toast, Alert, confirmAlert } from "@raycast/api";
 import path from "path";
+import os from "os";
 import { loadCategories, categoriesToFileTypes } from "./utils/categories";
-import { analyzeFolder, organizeFolder } from "./utils/file-organizer";
-import { pickFolder } from "./utils/folder-picker";
-import { chooseOrganizationMode } from "./utils/organization-mode";
+import { analyzeFolder, OrganizationMode, organizeFolder } from "./utils/file-organizer";
+import { OrganizationModePicker } from "./utils/organization-mode";
+import { formatSkippedSummary } from "./utils/organization-summary";
 
-export default async function main() {
+export default function Command() {
+  return <OrganizationModePicker onSelect={organizeDownloads} />;
+}
+
+async function organizeDownloads(mode: OrganizationMode) {
   try {
+    await closeMainWindow();
+
     // Load categories from storage
     const categories = await loadCategories();
     const fileTypes = categoriesToFileTypes(categories);
 
-    // Show folder picker dialog
-    const folderPath = await pickFolder("Select a folder to organize:");
-
-    if (!folderPath) {
-      await showToast({
-        style: Toast.Style.Success,
-        title: "Folder selection cancelled",
-      });
-      return;
-    }
-
-    const mode = await chooseOrganizationMode();
-
-    if (!mode) {
-      await showToast({
-        style: Toast.Style.Success,
-        title: "Organization cancelled",
-      });
-      return;
-    }
-
+    const downloadsPath = path.join(os.homedir(), "Downloads");
     // First, analyze files to get count
     const analysisToast = await showToast({
       style: Toast.Style.Animated,
-      title: "Scanning folder...",
-      message: `Analyzing files in ${path.basename(folderPath)}`,
+      title: "Scanning Downloads folder...",
     });
 
-    const analysisResult = analyzeFolder(folderPath, fileTypes, { mode });
+    const analysisResult = analyzeFolder(downloadsPath, fileTypes, { mode });
     analysisToast.hide();
 
     if (!analysisResult.success) {
@@ -57,7 +42,7 @@ export default async function main() {
       const skippedFolderCount = analysisResult.skipped_folders?.length || 0;
       await showToast({
         style: skippedFolderCount > 0 ? Toast.Style.Failure : Toast.Style.Success,
-        title: "Folder already clean",
+        title: "Downloads already clean",
         message: formatSkippedSummary(skippedProjectCount, skippedFolderCount) || "No files need to be sorted!",
       });
       return;
@@ -74,7 +59,7 @@ export default async function main() {
 
     const confirmed = await confirmAlert({
       title: `Sort ${analysisResult.total_files} files?`,
-      message: `Files in "${path.basename(folderPath)}" will be moved into folders:\n\n${categoryList}${formatSkippedSummary(
+      message: `Files will be moved into folders:\n\n${categoryList}${formatSkippedSummary(
         analysisResult.skipped_projects?.length || 0,
         analysisResult.skipped_folders?.length || 0,
         "\n\n",
@@ -101,20 +86,19 @@ export default async function main() {
     const sortingToast = await showToast({
       style: Toast.Style.Animated,
       title: "Sorting files...",
-      message: `Organizing files in ${path.basename(folderPath)}`,
     });
 
-    const sortResult = organizeFolder(folderPath, fileTypes, { mode });
+    const sortResult = organizeFolder(downloadsPath, fileTypes, { mode });
 
     if (!sortResult.success) {
       sortingToast.style = Toast.Style.Failure;
-      sortingToast.title = "❌ Organization failed";
+      sortingToast.title = "❌ Cleanup failed";
       sortingToast.message = sortResult.error || "Unknown error occurred";
       return;
     }
 
     sortingToast.style = Toast.Style.Success;
-    sortingToast.title = "✅ Folder organized!";
+    sortingToast.title = "✅ Downloads cleaned up!";
     sortingToast.message = `Sorted ${sortResult.total_moved || 0} files into ${
       sortResult.categories_created?.length || 0
     } folders${formatSkippedSummary(
@@ -122,42 +106,11 @@ export default async function main() {
       sortResult.skipped_folders?.length || 0,
       ". ",
     )}`;
-
-    // Offer to open the organized folder
-    const openFolder = await confirmAlert({
-      title: "Folder organized successfully!",
-      message: `${sortResult.total_moved} files have been sorted. Would you like to view the organized folder?`,
-      primaryAction: {
-        title: "Open Folder",
-        style: Alert.ActionStyle.Default,
-      },
-      dismissAction: {
-        title: "Done",
-        style: Alert.ActionStyle.Cancel,
-      },
-    });
-
-    if (openFolder) {
-      await showInFinder(folderPath);
-    }
   } catch (error) {
     await showToast({
       style: Toast.Style.Failure,
-      title: "❌ Organization failed",
+      title: "❌ Cleanup failed",
       message: error instanceof Error ? error.message : "Unknown error occurred",
     });
   }
-}
-
-function formatSkippedSummary(projectCount: number, folderCount: number, prefix = ""): string {
-  const parts: string[] = [];
-
-  if (projectCount > 0) {
-    parts.push(`Skipped ${projectCount} detected ${projectCount === 1 ? "project" : "projects"}`);
-  }
-  if (folderCount > 0) {
-    parts.push(`Could not scan ${folderCount} ${folderCount === 1 ? "folder" : "folders"}`);
-  }
-
-  return parts.length > 0 ? `${prefix}${parts.join(". ")}.` : "";
 }
